Adicao de Dockerfile para build de imagem e uwsgi para execucao da aplicacao
