+++ build_image.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+docker build -t tbd:latest .
