+++ docker-compose.yml
@@ -1,0 +1,31 @@
+version: '2.4'
+
+networks:
+  tbd_network:
+    ipam:
+      driver: default
+      config:
+        - subnet: "172.20.0.0/24"
+
+services:
+  mongo:
+    image: mongo
+    ports:
+      - "27017:27017"
+    restart: always
+    healthcheck:
+      test: ["CMD", "echo", "db.runCommand(\"ping\").ok", "|", "mongo", "mongo:27017/test", "--quiet"]
+      timeout: 20s
+      retries: 10
+    networks:
+      tbd_network:
+        aliases:
+          - mongo
+
+  application:
+    image: tbd:latest
+    ports:
+      - "8080:8080"
+    depends_on:
+      mongo:
+        condition: service_healthy
