+++ requirements.txt
@@ -1,3 +1,3 @@
-flask
-pymongo
-flask_pymongo
+flask==2.0.2
+pymongo==3.12.0
+uwsgi==2.0.19
