+++ src/wsgi.py
@@ -1,0 +1,3 @@
+from main import get_app
+
+app = get_app()
